MVS-607 Combine patient discharge with the submission of the vaccination record (#156)
* Combine submitVaccinationRecord and endEncounter into one step/action

* Fix merge issue

* fix merge issue
