+++ PointOfDispensingApp/src/components/VaccinationProceedComponent.vue
@@ -216,7 +216,10 @@
                   Back
                 </v-btn>
                 <v-spacer></v-spacer>
-                <v-btn color="primary" text @click="submitVaccinationRecord()">
+                <v-btn	
+                  color="primary"
+                  text
+                  @click="submitVaccinationRecord(); endEncounter();">
                   Submit
                 </v-btn>
               </v-card-actions>
@@ -245,7 +248,7 @@ export default {
     },
   },
   methods: {
-    onSuccess() {
+    onSuccessSubmitVaccinationRecord() {
       const vaccinationCompletePlayload = {
         lotNumber: this.lotNumber,
         expirationDate: this.expirationDate,
@@ -260,21 +263,39 @@ export default {
         notes: this.notes,
       };
 
+        //send data to Vuex
+        this.$store.dispatch("vaccinationComplete", vaccinationCompletePlayload);
+
+        //Advance to the Discharge page
+        this.$router.push("Discharge");
+
+        //Close the dialog
+        this.dialog = false;
+    },
+    onSuccessEndEncounter() {
+      //the following is sending dummy data until we have the API in place
+      const encounterResourcePayload = {
+        encounterStatus: "Finished",
+        encounterTimeStamp: new Date().toISOString(),
+      };
       //send data to Vuex
-      this.$store.dispatch("vaccinationComplete", vaccinationCompletePlayload);
-
-      //Advance to the Discharge page
-      this.$router.push("Discharge");
-
-      //Close the dialog
-      this.dialog = false;
+      this.$store.dispatch("patientDischarged", encounterResourcePayload);
     },
     submitVaccinationRecord() {
       brokerRequests.submitVaccination().then((response) => {
         if (response.data) {
-          this.onSuccess();
+          this.onSuccessSubmitVaccinationRecord();
         } else if (response.error) {
           alert("Vaccination record not submitted");
+        }
+      });
+    },
+    endEncounter() {
+      brokerRequests.discharge().then((response) => {
+        if (response.data) {
+          this.onSuccessEndEncounter();
+        } else if (response.error) {
+          alert("Discharge not successful");
         }
       });
     },
